owned_net_entity -> own file
